requirements.txt
anthropic>=0.84.0
rich>=13.0.0
python-dotenv>=1.0.0
requests>=2.31.0
beautifulsoup4>=4.12.0
lxml>=5.0.0
jinja2>=3.1.0
cvss>=2.6
dnspython>=2.6.0
weasyprint>=62.0
shodan>=1.31.0
